Refactor to forms components
